pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.115.14",
    "google-genai>=1.24.0",
    "google>=3.0.0",
    "langgraph>=0.5.1",
    "streamlit>=1.32.0",
    "pillow>=10.0.0",
    "pypdf>=5.7.0",
    "google-generativeai>=0.8.5",
    "chromadb>=1.0.15",
    "streamlit-autorefresh>=1.0.1",
]
